+++ src/perft.h
@@ -1,24 +1,9 @@
-#ifndef STRING_H_
-#define STRING_H_
-#include <string.h>
-#endif
+#pragma once
 
-#ifndef CONST_H_
-#define CONST_H_
-#include "board_constants.h"
-#endif
-
-#ifndef MOVES_H_
-#define MOVES_H_
 #include "moves_list.h"
-#endif
-
-#ifndef MOVEGEN_H_
-#define MOVEGEN_H_
 #include "movegen.h"
-#endif
+#include "board.h"
 
-#include "time.h"
 
 #ifndef U64
 #define U64 unsigned long long
@@ -59,37 +44,4 @@ static inline void perft_driver(int depth) {
     }
 }
 
-void perft_test(int depth) {
-
-    printf("\n  Performance test\n\n");
-
-    moves move_list[1];
-    generate_moves(move_list);
-
-    int start = get_time_ms();
-
-    for (int move_count = 0; move_count<move_list->count; move_count++) {
-        copy_board();
-
-        if (!make_move(move_list->moves[move_count], all_moves)) { 
-            continue;
-        }
-
-        long cumulative_nodes = nodes;
-
-        perft_driver(depth-1);
-
-        long old_nodes = nodes - cumulative_nodes;
-
-        take_back();
-
-        // print the test results
-        printf("  moves: ");
-        print_move(move_list->moves[move_count]);
-        printf(" nodes: %lld\n",old_nodes);
-    } 
-
-    printf("\n  Depth: %d\n", depth);
-    printf("  Nodes: %lld\n", nodes);
-    printf("  Time: %d ms\n", get_time_ms() - start);
-}
+void perft_test(int depth);
