+++ src/pre_calculated_tables.c
@@ -1,0 +1,40 @@
+#include "pre_calculated_tables.h"
+
+#include "magic_num_generator.h"
+#include "bit_manipulation.h"
+#include "masks.h"
+
+void init_leaper_attacks() {
+    for(int square = 0; square < 64; square++){
+        pawn_attacks[white][square] = mask_pawn_attacks(square,white);
+        pawn_attacks[black][square] = mask_pawn_attacks(square,black);
+
+        knight_attacks[square] = mask_knight_attacks(square);
+
+        king_attacks[square] = mask_king_attacks(square);
+    }
+}
+
+void init_sliders_attacks(int bishop) {
+    for (int square = 0; square < 64; square++) {
+        bishop_masks[square] = mask_bishop_attacks(square);
+        rook_masks[square] = mask_rook_attacks(square);
+
+        U64 attack_mask = bishop ? bishop_masks[square] : rook_masks[square];
+        int relevant_bits_count = count_bits(attack_mask);
+        int occupancy_indices = 1 << relevant_bits_count;
+
+        for (int index = 0; index < occupancy_indices; index++) {
+            if (bishop) {
+                U64 occupancy = set_occupancy(index, relevant_bits_count, attack_mask);
+                int magic_index = (occupancy * bishop_magic_numbers[square]) >> (64 - bishop_relevant_bits[square]);
+                bishop_attacks[square][magic_index] = bishop_attacks_on_the_fly(square,occupancy);
+            } else {
+                U64 occupancy = set_occupancy(index, relevant_bits_count, attack_mask);
+                int magic_index = (occupancy * rook_magic_numbers[square]) >> (64 - rook_relevant_bits[square]);
+                rook_attacks[square][magic_index] = rook_attacks_on_the_fly(square,occupancy);
+            }
+        }
+    }
+}
+
