+++ src/time.c
@@ -1,0 +1,7 @@
+#include <windows.h>
+
+#include "time.h"
+
+int get_time_ms() {
+    return GetTickCount();
+}
